Solution

+++ app/Http/Controllers/TaskController.php
@@ -16,10 +16,14 @@ class TaskController extends Controller
 
         $validated = $request->validate([
             'title' => 'required|string',
+            'due_date' => 'nullable|date',
+            'category_id' => 'nullable',
         ]);
 
         $task = new Task();
         $task->title = $validated['title'];
+        $task->due_date = $validated['due_date'];
+        $task->category_id = $validated['category_id'];
         $task->user()->associate($request->user());
         $task->save();
 
@@ -29,17 +33,20 @@ class TaskController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, $task)
+    public function update(Request $request, Task $task)
     {
-        $currentTask = Task::find($task);
 
-        $this->authorize('update', $currentTask);
+        // $currentTask = Task::find($task);
+      // dd($task);
+        $this->authorize('update', $task);
 
         $validated = $request->validate([
             'is_completed' => 'required|boolean',
         ]);
 
-        $currentTask->is_completed = $validated['is_completed'];
+        $task->is_completed = $validated['is_completed'];
+
+        $task->save();
 
         return back(303);
     }
@@ -47,8 +54,11 @@ class TaskController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy($task)
+    public function destroy(Task $task)
     {
+        // $currentTask = Task::find($task);
+
+        $this->authorize('delete', $task);
         $task->delete();
 
         return back(303);
